+++ package.json
@@ -79,7 +79,6 @@
     "react-dom": "^18.3.1",
     "react-outside-click-handler": "^1.3.0",
     "sharp": "^0.33.5",
-    "usehooks-ts": "^3.1.0",
     "webpack": "^5.94.0",
     "zod": "^3.23.8"
   },
